Fixes SNAP child support deduction applies incorrectly to gross vs. net income #2556
