+++ changelog_entry.yaml
@@ -2,3 +2,5 @@
   changes:
     fixed:
       - Updated Colorado SNAP parameters.
+      - Deduct child support from SNAP net income in all states.
+      - Deduct child support from SNAP gross income in certain states previously deducted from net income only.
