+++ policyengine_us/tests/policy/baseline/gov/usda/snap/income/gross/snap_child_support_gross_income_deduction.yaml
@@ -4,7 +4,7 @@
   input:
     child_support_expense: 100
   output:
-    snap_child_support_deduction: 0
+    snap_child_support_gross_income_deduction: 0
 
 - name: Alaska does deduct child support.
   period: 2022
@@ -13,4 +13,4 @@
     state_code: AK
     child_support_expense: 100
   output:
-    snap_child_support_deduction: 100
+    snap_child_support_gross_income_deduction: 100
