--- policyengine_us/tests/policy/baseline/gov/usda/snap/income/snap_gross_income.yaml
@@ -1,7 +1,0 @@
-- name: Gross income equals earned income plus unearned income.
-  period: 2020
-  input:
-    snap_earned_income: 1
-    snap_unearned_income: 2
-  output:
-    snap_gross_income: 3
