+++ policyengine_us/variables/gov/usda/snap/income/deductions/snap_child_support_deduction.py
@@ -12,10 +12,4 @@ class snap_child_support_deduction(Variable):
     definition_period = YEAR
     reference = "https://www.law.cornell.edu/uscode/text/7/2014#e_4"
 
-    def formula(spm_unit, period, parameters):
-        child_support = add(spm_unit, period, ["child_support_expense"])
-        state = spm_unit.household("state_code_str", period)
-        is_deductible = parameters(
-            period
-        ).gov.usda.snap.income.deductions.child_support[state]
-        return where(is_deductible, child_support, 0)
+    adds = ["child_support_expense"]
