+++ policyengine_us/variables/gov/usda/snap/income/gross/snap_gross_income.py
@@ -11,3 +11,6 @@ class snap_gross_income(Variable):
     unit = USD
 
     adds = ["snap_earned_income", "snap_unearned_income"]
+    # Only child support can be subtracted when computing gross income,
+    # and only in certain states.
+    subtracts = ["snap_child_support_gross_income_deduction"]
